Fixup application to be run from a folder
Prior to this commit, the application basically assumed it was running
on its own vhost. This is fine in many cases, but not all. This commit
changes the assumtpion to be '/xmas' but that should be rather easily
changed. Changes were required to the config.ru to ensure redirection
happened properly when coming from a reverse proxy such as Apache's
httpd. Additional changes were made to the views to ensure pathing was
relative and not absolute.

To change the folder name, you *should* only need to modify the BASEURL
constant in ChristmasTree.rb